Change file layout and add benchmarks
All the analyses, both the original and translated versions, now are in the main folder of the project. The python versions have to include the `skim.h` header in order to run.

A new folders has been added with the programs that need to be benchmarked. The files in this folder provide for templates for benchmarking that can be adjusted according to the resources and the number of runs wanted.